All other Api endpoints refactored. Bulk of the code removed to separate classes.
